+++ src/server.h
@@ -36,8 +36,6 @@ class EventFilter;
 namespace usc
 {
 class SessionSwitcher;
-class DMMessageHandler;
-class DMConnection;
 class Screen;
 class UnityDisplayService;
 class PowerButtonEventSink;
@@ -63,8 +61,6 @@ public:
     using mir::Server::the_compositor;
     using mir::Server::the_touch_visualizer;
 
-    virtual std::shared_ptr<DMMessageHandler> the_dm_message_handler();
-    virtual std::shared_ptr<DMConnection> the_dm_connection();
     virtual std::shared_ptr<Screen> the_screen();
     virtual std::shared_ptr<InputConfiguration> the_input_configuration();
     virtual std::shared_ptr<mir::input::EventFilter> the_screen_event_handler();
@@ -116,7 +112,6 @@ private:
 
     std::shared_ptr<SessionSwitcher> session_switcher;
 
-    mir::CachedPtr<DMConnection> dm_connection;
     mir::CachedPtr<Screen> screen;
     mir::CachedPtr<InputConfiguration> input_configuration;
     mir::CachedPtr<mir::input::EventFilter> screen_event_handler;
